+++ include/master_element/Tet4CVFEM.h
@@ -1,0 +1,180 @@
+/*------------------------------------------------------------------------*/
+/*  Copyright 2014 Sandia Corporation.                                    */
+/*  This software is released under the license detailed                  */
+/*  in the file, LICENSE, which is located in the top-level Nalu          */
+/*  directory structure                                                   */
+/*------------------------------------------------------------------------*/
+
+
+#ifndef Tet4CVFEM_h
+#define Tet4CVFEM_h
+
+#include<master_element/MasterElement.h>
+
+namespace sierra{
+namespace nalu{
+
+// Tet 4 subcontrol volume
+class TetSCV : public MasterElement
+{
+public:
+
+  TetSCV();
+  virtual ~TetSCV();
+
+  const int * ipNodeMap(int ordinal = 0);
+
+  void determinant(
+    SharedMemView<DoubleType**> coords,
+    SharedMemView<DoubleType*> volume);
+
+  void determinant(
+    const int nelem,
+    const double *coords,
+    double *areav,
+    double * error );
+
+  void shape_fcn(
+    double *shpfc);
+
+  void shifted_shape_fcn(
+    double *shpfc);
+  
+  void tet_shape_fcn(
+    const int &npts,
+    const double *par_coord, 
+    double* shape_fcn);
+};
+
+// Tet 4 subcontrol surface
+class TetSCS : public MasterElement
+{
+public:
+
+  TetSCS();
+  virtual ~TetSCS();
+
+  const int * ipNodeMap(int ordinal = 0);
+
+  virtual void determinant(
+    SharedMemView<DoubleType**>&coords,
+    SharedMemView<DoubleType**>&areav);
+
+  void determinant(
+    const int nelem,
+    const double *coords,
+    double *areav,
+    double * error );
+
+  void grad_op(
+    SharedMemView<DoubleType**>&coords,
+    SharedMemView<DoubleType***>&gradop,
+    SharedMemView<DoubleType***>&deriv);
+
+  void grad_op(
+    const int nelem,
+    const double *coords,
+    double *gradop,
+    double *deriv,
+    double *det_j,
+    double * error );
+
+  void shifted_grad_op(
+    SharedMemView<DoubleType**>&coords,
+    SharedMemView<DoubleType***>&gradop,
+    SharedMemView<DoubleType***>&deriv);
+
+  void shifted_grad_op(
+    const int nelem,
+    const double *coords,
+    double *gradop,
+    double *deriv,
+    double *det_j,
+    double * error );
+
+  void face_grad_op(
+    const int nelem,
+    const int face_ordinal,
+    const double *coords,
+    double *gradop,
+    double *det_j,
+    double * error );
+
+  void shifted_face_grad_op(
+    const int nelem,
+    const int face_ordinal,
+    const double *coords,
+    double *gradop,
+    double *det_j,
+    double * error );
+
+  void gij(
+    SharedMemView<DoubleType**>& coords,
+    SharedMemView<DoubleType***>& gupper,
+    SharedMemView<DoubleType***>& glower,
+    SharedMemView<DoubleType***>& deriv);
+
+  void gij(
+    const double *coords,
+    double *gupperij,
+    double *glowerij,
+    double *deriv);
+
+  const int * adjacentNodes();
+
+  void shape_fcn(
+    double *shpfc);
+
+  void shifted_shape_fcn(
+    double *shpfc);
+
+  void tet_shape_fcn(
+    const int &npts,
+    const double *par_coord,
+    double* shape_fcn);
+
+  int opposingNodes(
+    const int ordinal, const int node);
+
+  int opposingFace(
+    const int ordinal, const int node);
+
+  double isInElement(
+    const double *elemNodalCoord,
+    const double *pointCoord,
+    double *isoParCoord);
+
+  void interpolatePoint(
+    const int &nComp,
+    const double *isoParCoord,
+    const double *field,
+    double *result);
+
+  void general_shape_fcn(
+    const int numIp,
+    const double *isoParCoord,
+    double *shpfc);
+
+  void general_face_grad_op(
+    const int face_ordinal,
+    const double *isoParCoord,
+    const double *coords,
+    double *gradop,
+    double *det_j,
+    double * error );
+
+  void sidePcoords_to_elemPcoords(
+    const int & side_ordinal,
+    const int & npoints,
+    const double *side_pcoords,
+    double *elem_pcoords);
+
+  double parametric_distance(const std::vector<double> &x);
+
+  const int* side_node_ordinals(int sideOrdinal) final;
+};
+
+} // namespace nalu
+} // namespace Sierra
+
+#endif
