+++ src/master_element/MasterElementFactory.C
@@ -10,6 +10,7 @@
 
 #include <master_element/Hex8CVFEM.h>
 #include <master_element/Hex27CVFEM.h>
+#include <master_element/Tet4CVFEM.h>
 #include <master_element/MasterElementHO.h>
 
 #include <NaluEnv.h>
