+++ src/master_element/Tet4CVFEM.C
@@ -1,0 +1,1305 @@
+/*------------------------------------------------------------------------*/
+/*  Copyright 2014 Sandia Corporation.                                    */
+/*  This software is released under the license detailed                  */
+/*  in the file, LICENSE, which is located in the top-level Nalu          */
+/*  directory structure                                                   */
+/*------------------------------------------------------------------------*/
+
+
+#include <master_element/MasterElement.h>
+#include <master_element/MasterElementFunctions.h>
+#include <master_element/Tet4CVFEM.h>
+
+#include <AlgTraits.h>
+
+#include <NaluEnv.h>
+#include <FORTRAN_Proto.h>
+
+#include <stk_util/environment/ReportHandler.hpp>
+#include <stk_topology/topology.hpp>
+
+#include <iostream>
+
+#include <cmath>
+#include <limits>
+#include <array>
+#include <map>
+#include <memory>
+
+namespace sierra{
+namespace nalu{
+
+//--------------------------------------------------------------------------
+//-------- constructor -----------------------------------------------------
+//--------------------------------------------------------------------------
+TetSCV::TetSCV()
+  : MasterElement()
+{
+  nDim_ = 3;
+  nodesPerElement_ = 4;
+  numIntPoints_ = 4;
+
+  // define ip node mappings
+  ipNodeMap_.resize(4);
+  ipNodeMap_[0] = 0; ipNodeMap_[1] = 1; ipNodeMap_[2] = 2; ipNodeMap_[3] = 3;
+
+  // standard integration location
+  intgLoc_.resize(12);
+  const double seventeen96ths = 17.0/96.0;
+  const double fourfive96ths  = 45.0/96.0;
+  intgLoc_[0] = seventeen96ths; intgLoc_[1]  = seventeen96ths; intgLoc_[2]  = seventeen96ths; // vol 1
+  intgLoc_[3] = fourfive96ths;  intgLoc_[4]  = seventeen96ths; intgLoc_[5]  = seventeen96ths; // vol 2
+  intgLoc_[6] = seventeen96ths; intgLoc_[7]  = fourfive96ths;  intgLoc_[8]  = seventeen96ths; // vol 3
+  intgLoc_[9] = seventeen96ths; intgLoc_[10] = seventeen96ths; intgLoc_[11] = fourfive96ths;  // vol 4
+
+  // shifted
+  intgLocShift_.resize(12);
+  intgLocShift_[0] = 0.0; intgLocShift_[1]  = 0.0;  intgLocShift_[2] = 0.0;
+  intgLocShift_[3] = 1.0; intgLocShift_[4]  = 0.0;  intgLocShift_[5] = 0.0;
+  intgLocShift_[6] = 0.0; intgLocShift_[7]  = 1.0;  intgLocShift_[8] = 0.0;
+  intgLocShift_[9] = 0.0; intgLocShift_[10] = 0.0; intgLocShift_[11] = 1.0;
+}
+
+//--------------------------------------------------------------------------
+//-------- destructor ------------------------------------------------------
+//--------------------------------------------------------------------------
+TetSCV::~TetSCV()
+{
+  // does nothing
+}
+
+//--------------------------------------------------------------------------
+//-------- ipNodeMap -------------------------------------------------------
+//--------------------------------------------------------------------------
+const int *
+TetSCV::ipNodeMap(
+  int /*ordinal*/)
+{
+  // define scv->node mappings
+  return &ipNodeMap_[0];
+}
+
+void polyhedralVolumeByFaces(int ncoords, DoubleType volcoords[14][3], int ntriangles,
+                             const int triangleFaceTable[24][3], DoubleType& volume)
+{
+  DoubleType xface[3];
+
+  // initialize volume
+  volume = 0.0;
+
+  //loop over each triangular facet
+  for(int itriangle=0; itriangle<ntriangles; ++itriangle) {
+    int ip = triangleFaceTable[itriangle][0];
+    int iq = triangleFaceTable[itriangle][1];
+    int ir = triangleFaceTable[itriangle][2];
+    // set spatial coordinate of integration point
+    for(int k=0; k<3; ++k) {
+      xface[k] = volcoords[ip][k] + volcoords[iq][k] + volcoords[ir][k];
+    }
+    // calculate contribution of triangular face to volume
+    volume +=
+        xface[0]*( ( volcoords[iq][1]-volcoords[ip][1] )*
+                   ( volcoords[ir][2]-volcoords[ip][2] )
+                 - ( volcoords[ir][1]-volcoords[ip][1] )*
+                   ( volcoords[iq][2]-volcoords[ip][2] ) )
+      - xface[1]*( ( volcoords[iq][0]-volcoords[ip][0] )*
+                   ( volcoords[ir][2]-volcoords[ip][2] )
+                 - ( volcoords[ir][0]-volcoords[ip][0] )*
+                   ( volcoords[iq][2]-volcoords[ip][2] ) )
+      + xface[2]*( ( volcoords[iq][0]-volcoords[ip][0] )*
+                   ( volcoords[ir][1]-volcoords[ip][1] )
+                 - ( volcoords[ir][0]-volcoords[ip][0] )*
+                   ( volcoords[iq][1]-volcoords[ip][1] ) );
+  }
+  volume /= 18.0;
+}
+
+void hexVolumeByTriangleFacets(const DoubleType volcoords[8][3], DoubleType& volume)
+{
+  const int triangularFacetTable[24][3] = {
+    {0, 8, 1},
+    {8, 2, 1},
+    {3, 2, 8},
+    {3, 8, 0},
+    {6, 9, 5},
+    {7, 9, 6},
+    {4, 9, 7},
+    {4, 5, 9},
+    {10, 0, 1},
+    {5, 10, 1},
+    {4, 10, 5},
+    {4, 0, 10}, 
+    {7, 6, 11}, 
+    {6, 2, 11}, 
+    {2, 3, 11}, 
+    {3, 7, 11}, 
+    {6, 12, 2},
+    {5, 12, 6},
+    {5, 1, 12}, 
+    {1, 2, 12}, 
+    {0, 4, 13}, 
+    {4, 7, 13}, 
+    {7, 3, 13}, 
+    {3, 0, 13}
+  };
+
+  DoubleType coords[14][3];
+  const double one4th = 0.25;
+  const int dim[3] = {0, 1, 2};
+
+  // the first coordinates are equivalent to vertices
+  for(int j=0; j<8; ++j) {
+    for(int k : dim) {
+      coords[j][k] = volcoords[j][k];
+    }
+  }
+  // now we add the face midpoints
+  for(int k : dim) {
+    coords[8][k] = one4th*( volcoords[0][k] + volcoords[1][k] + volcoords[2][k] + volcoords[3][k] );
+  }
+  for(int k : dim) {
+    coords[9][k] = one4th*( volcoords[4][k] + volcoords[5][k] + volcoords[6][k] + volcoords[7][k] );
+  }
+  for(int k : dim) {
+    coords[10][k] = one4th*( volcoords[0][k] + volcoords[1][k] + volcoords[5][k] + volcoords[4][k] );
+  }
+  for(int k : dim) {
+    coords[11][k] = one4th*( volcoords[3][k] + volcoords[2][k] + volcoords[6][k] + volcoords[7][k] );
+  }
+  for(int k : dim) {
+    coords[12][k] = one4th*( volcoords[1][k] + volcoords[2][k] + volcoords[6][k] + volcoords[5][k] );
+  }
+  for(int k : dim) {
+    coords[13][k] = one4th*( volcoords[0][k] + volcoords[3][k] + volcoords[7][k] + volcoords[4][k] );
+  }
+
+  // number of vertices for the equivalent tetrahedron
+  int ncoords = 14;
+  // number of triangular facets
+  int ntriangles = 24;
+
+  // calculate the volume using the triangular facets
+  polyhedralVolumeByFaces(ncoords, coords, ntriangles, triangularFacetTable, volume);
+}
+
+//--------------------------------------------------------------------------
+//-------- determinant -----------------------------------------------------
+//--------------------------------------------------------------------------
+void TetSCV::determinant(
+    SharedMemView<DoubleType**> coordel,
+    SharedMemView<DoubleType*> volume)
+{
+  const int tetSubcontrolNodeTable[4][8] = {
+    {0, 4, 7, 6, 11, 13, 14, 12},
+    {1, 5, 7, 4, 9, 10, 14, 13},
+    {2, 6, 7, 5, 8, 12, 14, 10},
+    {3, 9, 13, 11, 8, 10, 14, 12}
+  };
+
+  const double half = 0.5;
+  const double one3rd = 1.0/3.0;
+  DoubleType coords[15][3];
+  DoubleType ehexcoords[8][3];
+  const int dim[3] = {0, 1, 2};
+
+  // element vertices
+  for(int j=0; j<4; ++j) {
+    for(int k : dim) {
+      coords[j][k] = coordel(j,k);
+    }
+  }
+
+  // face 1 (tri)
+
+  // edge midpoints
+  for(int k : dim) {
+    coords[4][k] = half*(coordel(0,k) + coordel(1,k));
+  }
+  for(int k : dim) {
+    coords[5][k]= half*(coordel(1,k) + coordel(2,k));
+  }
+  for(int k : dim) {
+    coords[6][k] = half*(coordel(2,k) + coordel(0,k));
+  }
+
+  // face mipdoint
+  for(int k : dim) {
+    coords[7][k] = one3rd*(coordel(0,k) + coordel(1,k) + coordel(2,k));
+  }
+
+  // face 2 (tri)
+
+  // edge midpoints
+  for(int k : dim) {
+    coords[8][k] = half*(coordel(2,k) + coordel(3,k));
+  }
+  for(int k : dim) {
+    coords[9][k] = half*(coordel(3,k) + coordel(1,k));
+  }
+
+  // face midpoint
+  for(int k : dim) {
+    coords[10][k] = one3rd*(coordel(1,k) + coordel(2,k) + coordel(3,k));
+  }
+
+  // face 3 (tri)
+
+  // edge midpoint
+  for(int k : dim) {
+    coords[11][k] = half*(coordel(0,k) + coordel(3,k));
+  }
+
+  // face midpoint
+  for(int k : dim) {
+    coords[12][k] = one3rd*(coordel(0,k) + coordel(2,k) + coordel(3,k));
+  }
+
+  // face 4 (tri)
+
+  // face midpoint
+  for(int k : dim) {
+    coords[13][k] = one3rd*(coordel(0,k) + coordel(1,k) + coordel(3,k));
+  }
+
+  // element centroid
+  for(int k : dim) {
+    coords[14][k] = 0.0;
+  }
+  for(int j=0; j<nodesPerElement_; ++j) {
+    for(int k : dim) {
+      coords[14][k] = coords[14][k] + 0.25*coordel(j,k);
+    }
+  }
+
+  // loop over subcontrol volumes
+  for(int icv=0; icv<numIntPoints_; ++icv) {
+    // loop over nodes of scv
+    for(int inode=0; inode<8; ++inode) {
+      // define scv coordinates using node table
+      for(int k : dim) {
+        ehexcoords[inode][k] = coords[tetSubcontrolNodeTable[icv][inode]][k];
+      }
+    }
+    // compute volume using an equivalent polyhedron
+    DoubleType voltmp = 0.0;
+    hexVolumeByTriangleFacets(ehexcoords,voltmp);
+    volume(icv) = voltmp;
+    // check for negative volume
+    //ThrowAssertMsg( volume(icv) < 0.0, "ERROR in TetSCV::determinant, negative volume.");
+  }
+}
+
+void TetSCV::determinant(
+  const int nelem,
+  const double *coords,
+  double *volume,
+  double *error)
+{
+  int lerr = 0;
+
+  SIERRA_FORTRAN(tet_scv_det)
+    ( &nelem, &nodesPerElement_, &numIntPoints_, coords,
+      volume, error, &lerr );
+}
+
+//--------------------------------------------------------------------------
+//-------- shape_fcn -------------------------------------------------------
+//--------------------------------------------------------------------------
+void
+TetSCV::shape_fcn(double *shpfc)
+{
+  tet_shape_fcn(numIntPoints_, &intgLoc_[0], shpfc);
+}
+
+//--------------------------------------------------------------------------
+//-------- shifted_shape_fcn -----------------------------------------------
+//--------------------------------------------------------------------------
+void
+TetSCV::shifted_shape_fcn(double *shpfc)
+{
+  tet_shape_fcn(numIntPoints_, &intgLocShift_[0], shpfc);
+}
+
+//--------------------------------------------------------------------------
+//-------- tet_shape_fcn ---------------------------------------------------
+//--------------------------------------------------------------------------
+void
+TetSCV::tet_shape_fcn(
+  const int  &npts,
+  const double *par_coord, 
+  double *shape_fcn)
+{
+  for (int j = 0; j < npts; ++j ) {
+    const int fourj = 4*j;
+    const int k = 3*j;
+    const double xi = par_coord[k];
+    const double eta = par_coord[k+1];
+    const double zeta = par_coord[k+2];
+    shape_fcn[fourj] = 1.0 - xi - eta - zeta;
+    shape_fcn[1 + fourj] = xi;
+    shape_fcn[2 + fourj] = eta;
+    shape_fcn[3 + fourj] = zeta;
+  }
+}
+
+//--------------------------------------------------------------------------
+//-------- constructor -----------------------------------------------------
+//--------------------------------------------------------------------------
+TetSCS::TetSCS()
+  : MasterElement()
+{
+  nDim_ = 3;
+  nodesPerElement_ = 4;
+  numIntPoints_ = 6;
+
+  // define L/R mappings
+  lrscv_.resize(12);
+  lrscv_[0]  = 0; lrscv_[1]  = 1;
+  lrscv_[2]  = 1; lrscv_[3]  = 2;
+  lrscv_[4]  = 0; lrscv_[5]  = 2;
+  lrscv_[6]  = 0; lrscv_[7]  = 3;
+  lrscv_[8]  = 1; lrscv_[9]  = 3;
+  lrscv_[10] = 2; lrscv_[11] = 3;
+
+  // define opposing node
+  oppNode_.resize(12);
+  // face 0
+  oppNode_[0] = 2; oppNode_[1] = 2;  oppNode_[2] = 2;
+  // face 1
+  oppNode_[3] = 0; oppNode_[4] = 0;  oppNode_[5] = 0;
+  // face 2
+  oppNode_[6] = 1; oppNode_[7] = 1;  oppNode_[8] = 1;
+  // face 3
+  oppNode_[9] = 3; oppNode_[10] = 3; oppNode_[11] = 3;
+
+  // define opposing face
+  oppFace_.resize(12);
+  // face 0
+  oppFace_[0]  = 2; oppFace_[1] = 1;  oppFace_[2] = 5;
+  // face 1
+  oppFace_[3]  = 0; oppFace_[4] = 2;  oppFace_[5] = 3;
+  // face 2
+  oppFace_[6]  = 0; oppFace_[7] = 4;  oppFace_[8] = 1;
+  // face 3
+  oppFace_[9]  = 3; oppFace_[10] = 5; oppFace_[11] = 4;
+
+  // standard integration location
+  intgLoc_.resize(18);
+  const double seventeen48ths = 17.0/48.0;
+  const double seven48ths = 7.0/48.0;
+  intgLoc_[0]  =  seventeen48ths; intgLoc_[1]  = seven48ths;     intgLoc_[2]  = seven48ths; // surf 1    1->2
+  intgLoc_[3]  =  seventeen48ths; intgLoc_[4]  = seventeen48ths; intgLoc_[5]  = seven48ths; // surf 2    2->3
+  intgLoc_[6]  =  seven48ths;     intgLoc_[7]  = seventeen48ths; intgLoc_[8]  = seven48ths; // surf 3    1->3
+  intgLoc_[9]  =  seven48ths ;    intgLoc_[10] = seven48ths;     intgLoc_[11] = seventeen48ths; // surf 4    1->4
+  intgLoc_[12] =  seventeen48ths; intgLoc_[13] = seven48ths;     intgLoc_[14] = seventeen48ths; // surf 5    2->4
+  intgLoc_[15] =  seven48ths;     intgLoc_[16] = seventeen48ths; intgLoc_[17] = seventeen48ths; // surf 6    3->4
+
+  // shifted
+  intgLocShift_.resize(36);
+  intgLocShift_[0]  =  0.50; intgLocShift_[1]  =  0.00; intgLocShift_[2]  =  0.00; // surf 1    1->2
+  intgLocShift_[3]  =  0.50; intgLocShift_[4]  =  0.50; intgLocShift_[5]  =  0.00; // surf 2    2->3
+  intgLocShift_[6]  =  0.00; intgLocShift_[7]  =  0.50; intgLocShift_[8]  =  0.00; // surf 3    1->3
+  intgLocShift_[9]  =  0.00; intgLocShift_[10] =  0.00; intgLocShift_[11] =  0.50; // surf 4    1->4
+  intgLocShift_[12] =  0.50; intgLocShift_[13] =  0.00; intgLocShift_[14] =  0.50; // surf 5    2->4
+  intgLocShift_[15] =  0.00; intgLocShift_[16] =  0.50; intgLocShift_[17] =  0.50; // surf 6    3->4
+
+  // exposed face
+  intgExpFace_.resize(36);
+  const double five24ths = 5.0/24.0;
+  const double seven12ths = 7.0/12.0;
+  // face 0; nodes 0,1,3: scs 0, 1, 2
+  intgExpFace_[0]  = five24ths;  intgExpFace_[1]  =  0.00; intgExpFace_[2]  = five24ths;
+  intgExpFace_[3]  = seven12ths; intgExpFace_[4]  =  0.00; intgExpFace_[5]  = five24ths;
+  intgExpFace_[6]  = five24ths;  intgExpFace_[7]  =  0.00; intgExpFace_[8]  = seven12ths;
+  // face 1; nodes 1,2,3; scs 0, 1, 2
+  intgExpFace_[9]  = seven12ths; intgExpFace_[10] = five24ths;  intgExpFace_[11] = five24ths;
+  intgExpFace_[12] = five24ths;  intgExpFace_[13] = seven12ths; intgExpFace_[14] = five24ths;
+  intgExpFace_[15] = five24ths;  intgExpFace_[16] = five24ths;  intgExpFace_[17] = seven12ths;
+  // face 2; nodes 0,3,2; scs 0, 1, 2
+  intgExpFace_[18] =  0.00;      intgExpFace_[19] = five24ths;  intgExpFace_[20] = five24ths;
+  intgExpFace_[21] =  0.00;      intgExpFace_[22] = five24ths;  intgExpFace_[23] = seven12ths;
+  intgExpFace_[24] =  0.00;      intgExpFace_[25] = seven12ths; intgExpFace_[26] = five24ths;
+  //face 3; nodes 0, 2, 1; scs 0, 1, 2
+  intgExpFace_[27] = five24ths;  intgExpFace_[28] = five24ths;  intgExpFace_[29] =  0.00;
+  intgExpFace_[30] = seven12ths; intgExpFace_[31] = five24ths;  intgExpFace_[32] =  0.00;
+  intgExpFace_[33] = five24ths;  intgExpFace_[34] = seven12ths; intgExpFace_[35] =  0.00;
+
+  // boundary integration point ip node mapping (ip on an ordinal to local node number)
+  ipNodeMap_.resize(12); // 3 ips * 4 faces
+  // face 0;
+  ipNodeMap_[0] = 0;  ipNodeMap_[1] = 1;  ipNodeMap_[2] = 3; 
+  // face 1; 
+  ipNodeMap_[3] = 1;  ipNodeMap_[4] = 2;  ipNodeMap_[5] = 3; 
+  // face 2;
+  ipNodeMap_[6] = 0;  ipNodeMap_[7] = 3;  ipNodeMap_[8] = 2;  
+  // face 3;
+  ipNodeMap_[9] = 0; ipNodeMap_[10] = 2; ipNodeMap_[11] = 1;
+
+  sideNodeOrdinals_ = {
+      0, 1, 3, //ordinal 0
+      1, 2, 3, //ordinal 1
+      0, 3, 2, //ordinal 2
+      0, 2, 1  //ordinal 3
+  };
+
+  std::vector<std::vector<double>> nodeLocations =
+  {
+      {0,0,0}, {1,0,0}, {0,1,0}, {0,0,1}
+  };
+
+  intgExpFaceShift_.resize(3*3*4);
+  int index = 0;
+  stk::topology topo = stk::topology::TET_4;
+  for (unsigned k = 0; k < topo.num_sides(); ++k) {
+    stk::topology side_topo = topo.side_topology(k);
+    const int* ordinals = side_node_ordinals(k);
+    for (unsigned n = 0; n < side_topo.num_nodes(); ++n) {
+      intgExpFaceShift_[3*index + 0] = nodeLocations[ordinals[n]][0];
+      intgExpFaceShift_[3*index + 1] = nodeLocations[ordinals[n]][1];
+      intgExpFaceShift_[3*index + 2] = nodeLocations[ordinals[n]][2];
+      ++index;
+    }
+  }
+}
+
+//--------------------------------------------------------------------------
+//-------- destructor ------------------------------------------------------
+//--------------------------------------------------------------------------
+TetSCS::~TetSCS()
+{
+  // does nothing
+}
+
+//--------------------------------------------------------------------------
+//-------- ipNodeMap -------------------------------------------------------
+//--------------------------------------------------------------------------
+const int *
+TetSCS::ipNodeMap(
+  int ordinal)
+{
+  // define ip->node mappings for each face (ordinal); 
+  return &ipNodeMap_[ordinal*3];
+}
+
+//--------------------------------------------------------------------------
+//-------- side_node_ordinals ----------------------------------------------
+//--------------------------------------------------------------------------
+const int *
+TetSCS::side_node_ordinals(
+  int ordinal)
+{
+  // define face_ordinal->node_ordinal mappings for each face (ordinal);
+  return &sideNodeOrdinals_[ordinal*3];
+}
+
+void quadAreaByTriFacets(DoubleType areacoords[4][3], DoubleType* area)
+{
+  int triFacetTable[4][3] = {
+    {4, 0, 1},
+    {4, 1, 2},
+    {4, 2, 3},
+    {4, 3, 0}
+  };
+  const double one4th = 1.0/4.0;
+  const double half = 1.0/2.0;
+  const int dim[] = {0, 1, 2};
+  DoubleType xmid[3];
+  DoubleType r1[3];
+  DoubleType r2[3];
+
+  for(int k : dim) {
+    xmid[k] = one4th*(areacoords[0][k] + areacoords[1][k] + areacoords[2][k] + areacoords[3][k]);
+  }
+
+  const int ntriangles = 4;
+  for(int k : dim) {
+    area[k] = 0.0;
+    r2[k] = areacoords[0][k] - xmid[k];
+  }
+
+  for(int itriangle=0; itriangle<ntriangles; ++itriangle) {
+    int iq = triFacetTable[itriangle][2];
+    //construct vectors with common beginning point
+//    r1 = r2;
+    for(int k : dim) {
+      r1[k] = r2[k];
+      r2[k] = areacoords[iq][k] - xmid[k];
+    }
+    //cross product is twice the area vector
+    //triFacetTable should be constructed such
+    //that these vectors have the same convention (right hand rule)
+    area[0] += r1[1]*r2[2] - r2[1]*r1[2];
+    area[1] += r1[2]*r2[0] - r2[2]*r1[0];
+    area[2] += r1[0]*r2[1] - r2[0]*r1[1];
+  }
+
+  for(int k : dim) {
+    area[k] = half*area[k];
+  }
+}
+
+//--------------------------------------------------------------------------
+//-------- determinant -----------------------------------------------------
+//--------------------------------------------------------------------------
+void TetSCS::determinant(
+    SharedMemView<DoubleType**>& coordel,
+    SharedMemView<DoubleType**>&areav)
+{
+  int tetEdgeFacetTable[6][4] = {
+    {4, 7, 14, 13},
+    {7, 14, 10, 5},
+    {6, 12, 14, 7},
+    {11, 13, 14, 12},
+    {13, 9, 10, 14},
+    {10, 8, 12, 14}
+  };
+
+  const int npe = nodesPerElement_;
+  const int nscs = numIntPoints_;
+  const double half = 0.5;
+  const double one3rd = 1.0/3.0;
+  const double one4th = 1.0/4.0;
+  const int dim[] = {0, 1, 2};
+  DoubleType coords[15][3];
+  DoubleType scscoords[4][3];
+
+  //element vertices
+  for(int j=0; j<4; ++j) {
+    for(int k : dim) {
+      coords[j][k] = coordel(j,k);
+    }
+  }
+
+  //face 1 (tri)
+  //
+  //edge midpoints
+  for(int k : dim) {
+    coords[4][k] = half*(coordel(0,k) + coordel(1,k));
+  }
+  for(int k : dim) {
+    coords[5][k] = half*(coordel(1,k) + coordel(2,k));
+  }
+  for(int k : dim) {
+    coords[6][k] = half*(coordel(2,k) + coordel(0,k));
+  }
+
+  //face midpoint
+  for(int k : dim) {
+    coords[7][k] = one3rd*(coordel(0,k) + coordel(1,k) + coordel(2,k));
+  }
+
+  //face 2 (tri)
+  //
+  //edge midpoints
+  for(int k : dim) {
+    coords[8][k] = half*(coordel(2,k) + coordel(3,k));
+  }
+  for(int k : dim) {
+    coords[9][k] = half*(coordel(3,k) + coordel(1,k));
+  }
+
+  //face midpoint
+  for(int k : dim) {
+    coords[10][k] = one3rd*(coordel(1,k) + coordel(2,k) + coordel(3,k));
+  }
+
+  //face 3 (tri)
+  //
+  //edge midpoint
+  for(int k : dim) {
+    coords[11][k] = half*(coordel(0,k) + coordel(3,k));
+  }
+
+  //face midpoint
+  for(int k : dim) {
+    coords[12][k] = one3rd*(coordel(0,k) + coordel(2,k) + coordel(3,k));
+  }
+
+  //face 4 (tri)
+  //
+  //face midpoint
+  for(int k : dim) {
+    coords[13][k] = one3rd*(coordel(0,k) + coordel(1,k) + coordel(3,k));
+  }
+
+  //element centroid
+  for(int k : dim) {
+    coords[14][k] = 0.0;
+  }
+  for(int j=0; j<npe; ++j) {
+    for(int k : dim) {
+      coords[14][k] += one4th*coordel(j,k);
+    }
+  }
+
+  //loop over subcontrol surface
+  for(int ics=0; ics<nscs; ++ics) {
+    //loop over nodes of scs
+    for(int inode=0; inode<4; ++inode) {
+      int itrianglenode = tetEdgeFacetTable[ics][inode];
+      for(int k : dim) {
+        scscoords[inode][k] = coords[itrianglenode][k];
+      }
+    }
+    quadAreaByTriFacets(scscoords, &areav(ics,0));
+  }
+}
+
+void TetSCS::determinant(
+  const int nelem,
+  const double *coords,
+  double *areav,
+  double *error)
+{
+  SIERRA_FORTRAN(tet_scs_det)
+    ( &nelem, &nodesPerElement_, &numIntPoints_, coords, areav );
+
+  // all is always well; no error checking
+  *error = 0;
+}
+
+void tet_deriv(SharedMemView<DoubleType***>& deriv)
+{
+  for(size_t j=0; j<deriv.dimension(0); ++j) {
+    deriv(j,0,0) = -1.0;
+    deriv(j,0,1) = -1.0;
+    deriv(j,0,2) = -1.0;
+              
+    deriv(j,1,0) = 1.0;
+    deriv(j,1,1) = 0.0;
+    deriv(j,1,2) = 0.0;
+              
+    deriv(j,2,0) = 0.0;
+    deriv(j,2,1) = 1.0;
+    deriv(j,2,2) = 0.0;
+              
+    deriv(j,3,0) = 0.0;
+    deriv(j,3,1) = 0.0;
+    deriv(j,3,2) = 1.0;
+  }
+}
+
+void tet_grad_op(SharedMemView<DoubleType***>& deriv,
+                 SharedMemView<DoubleType**>& coordel,
+                 SharedMemView<DoubleType***>& gradop)
+{
+  size_t nint = deriv.dimension(0);
+  size_t npe = deriv.dimension(1);
+
+  DoubleType dx_ds1, dx_ds2, dx_ds3;
+  DoubleType dy_ds1, dy_ds2, dy_ds3;
+  DoubleType dz_ds1, dz_ds2, dz_ds3;
+  DoubleType ds1_dx, ds1_dy, ds1_dz;
+  DoubleType ds2_dx, ds2_dy, ds2_dz;
+  DoubleType ds3_dx, ds3_dy, ds3_dz;
+
+  double realmin = 2.2250738585072014e-308;
+
+  for(size_t ki=0; ki<nint; ++ki) {
+    dx_ds1 = 0.0;
+    dx_ds2 = 0.0;
+    dx_ds3 = 0.0;
+    dy_ds1 = 0.0;
+    dy_ds2 = 0.0;
+    dy_ds3 = 0.0;
+    dz_ds1 = 0.0;
+    dz_ds2 = 0.0;
+    dz_ds3 = 0.0;
+//
+// calculate the jacobian at the integration station -
+//
+    for(size_t kn=0; kn<npe; ++kn) {
+      dx_ds1 = dx_ds1+deriv(ki,kn,0)*coordel(kn,0);
+      dx_ds2 = dx_ds2+deriv(ki,kn,1)*coordel(kn,0);
+      dx_ds3 = dx_ds3+deriv(ki,kn,2)*coordel(kn,0);
+
+      dy_ds1 = dy_ds1+deriv(ki,kn,0)*coordel(kn,1);
+      dy_ds2 = dy_ds2+deriv(ki,kn,1)*coordel(kn,1);
+      dy_ds3 = dy_ds3+deriv(ki,kn,2)*coordel(kn,1);
+
+      dz_ds1 = dz_ds1+deriv(ki,kn,0)*coordel(kn,2);
+      dz_ds2 = dz_ds2+deriv(ki,kn,1)*coordel(kn,2);
+      dz_ds3 = dz_ds3+deriv(ki,kn,2)*coordel(kn,2);
+    }
+
+//
+// calculate the determinate of the jacobian at the integration station
+//
+    DoubleType det_j = dx_ds1*( dy_ds2*dz_ds3 - dz_ds2*dy_ds3 )
+                     + dy_ds1*( dz_ds2*dx_ds3 - dx_ds2*dz_ds3 )
+                     + dz_ds1*( dx_ds2*dy_ds3 - dy_ds2*dx_ds3 );
+
+//
+// protect against a negative or small value for the determinate of the
+// jacobian. The value of real_min represents
+// the smallest Real value (based upon the precision set for this
+// compilation) which the machine can represent -
+//
+    DoubleType test = stk::math::if_then_else(det_j > 1.e+6*realmin, det_j, 1.0);
+    DoubleType denom = 1.0/test;
+//
+// compute the gradient operators at the integration station -
+//
+    ds1_dx = denom*(dy_ds2*dz_ds3 - dz_ds2*dy_ds3);
+    ds2_dx = denom*(dz_ds1*dy_ds3 - dy_ds1*dz_ds3);
+    ds3_dx = denom*(dy_ds1*dz_ds2 - dz_ds1*dy_ds2);
+
+    ds1_dy = denom*(dz_ds2*dx_ds3 - dx_ds2*dz_ds3);
+    ds2_dy = denom*(dx_ds1*dz_ds3 - dz_ds1*dx_ds3);
+    ds3_dy = denom*(dz_ds1*dx_ds2 - dx_ds1*dz_ds2);
+
+    ds1_dz = denom*(dx_ds2*dy_ds3 - dy_ds2*dx_ds3);
+    ds2_dz = denom*(dy_ds1*dx_ds3 - dx_ds1*dy_ds3);
+    ds3_dz = denom*(dx_ds1*dy_ds2 - dy_ds1*dx_ds2);
+
+    for(size_t kn=0; kn<npe; ++kn) {
+      gradop(ki,kn,0) =
+          deriv(ki,kn,0)*ds1_dx
+        + deriv(ki,kn,1)*ds2_dx
+        + deriv(ki,kn,2)*ds3_dx;
+
+      gradop(ki,kn,1) =
+          deriv(ki,kn,0)*ds1_dy
+        + deriv(ki,kn,1)*ds2_dy
+        + deriv(ki,kn,2)*ds3_dy;
+
+      gradop(ki,kn,2) =
+          deriv(ki,kn,0)*ds1_dz
+        + deriv(ki,kn,1)*ds2_dz
+        + deriv(ki,kn,2)*ds3_dz;
+    }
+  }
+}
+
+//--------------------------------------------------------------------------
+//-------- grad_op ---------------------------------------------------------
+//--------------------------------------------------------------------------
+void TetSCS::grad_op(
+    SharedMemView<DoubleType**>&coords,
+    SharedMemView<DoubleType***>&gradop,
+    SharedMemView<DoubleType***>&deriv)
+{
+  tet_deriv(deriv);
+
+  tet_grad_op(deriv, coords, gradop);
+}
+
+void TetSCS::grad_op(
+  const int nelem,
+  const double *coords,
+  double *gradop,
+  double *deriv,
+  double *det_j,
+  double *error)
+{
+  int lerr = 0;
+
+  SIERRA_FORTRAN(tet_derivative)
+    ( &numIntPoints_, deriv );
+  
+  SIERRA_FORTRAN(tet_gradient_operator)
+    ( &nelem,
+      &nodesPerElement_,
+      &numIntPoints_,
+      deriv,
+      coords, gradop, det_j, error, &lerr );
+
+  if ( lerr )
+    std::cout << "sorry, negative TetSCS volume.." << std::endl;  
+}
+
+//--------------------------------------------------------------------------
+//-------- shifted_grad_op -------------------------------------------------
+//--------------------------------------------------------------------------
+void TetSCS::shifted_grad_op(
+    SharedMemView<DoubleType**>&coords,
+    SharedMemView<DoubleType***>&gradop,
+    SharedMemView<DoubleType***>&deriv)
+{
+  tet_deriv(deriv);
+
+  tet_grad_op(deriv, coords, gradop);
+}
+
+void TetSCS::shifted_grad_op(
+  const int nelem,
+  const double *coords,
+  double *gradop,
+  double *deriv,
+  double *det_j,
+  double *error)
+{
+  int lerr = 0;
+
+  SIERRA_FORTRAN(tet_derivative)
+    ( &numIntPoints_, deriv );
+
+  SIERRA_FORTRAN(tet_gradient_operator)
+    ( &nelem,
+      &nodesPerElement_,
+      &numIntPoints_,
+      deriv,
+      coords, gradop, det_j, error, &lerr );
+
+  if ( lerr )
+    std::cout << "sorry, negative TetSCS volume.." << std::endl;
+}
+
+//--------------------------------------------------------------------------
+//-------- face_grad_op ----------------------------------------------------
+//--------------------------------------------------------------------------
+void TetSCS::face_grad_op(
+  const int nelem,
+  const int /*face_ordinal*/,
+  const double *coords,
+  double *gradop,
+  double *det_j,
+  double *error)
+{
+  int lerr = 0;
+  int npf = 3;
+
+  const int nface = 1;
+  double dpsi[12];
+
+  for ( int n=0; n<nelem; n++ ) {
+
+    for ( int k=0; k<npf; k++ ) {
+
+      // derivatives are constant
+      SIERRA_FORTRAN(tet_derivative)
+        ( &nface, dpsi );
+
+      SIERRA_FORTRAN(tet_gradient_operator)
+        ( &nface,
+          &nodesPerElement_,
+          &nface,
+          dpsi,
+          &coords[12*n], &gradop[k*nelem*12+n*12], &det_j[npf*n+k], error, &lerr );
+
+      if ( lerr )
+        std::cout << "sorry, issue with face_grad_op.." << std::endl;
+
+    }
+  }
+}
+
+//--------------------------------------------------------------------------
+//-------- shifted_face_grad_op --------------------------------------------
+//--------------------------------------------------------------------------
+void TetSCS::shifted_face_grad_op(
+  const int nelem,
+  const int /*face_ordinal*/,
+  const double *coords,
+  double *gradop,
+  double *det_j,
+  double *error)
+{
+  // no difference for regular face_grad_op
+
+  int lerr = 0;
+  int npf = 3;
+
+  const int nface = 1;
+  double dpsi[12];
+
+  for ( int n=0; n<nelem; n++ ) {
+
+    for ( int k=0; k<npf; k++ ) {
+
+      // derivatives are constant
+      SIERRA_FORTRAN(tet_derivative)
+        ( &nface, dpsi );
+
+      SIERRA_FORTRAN(tet_gradient_operator)
+        ( &nface,
+          &nodesPerElement_,
+          &nface,
+          dpsi,
+          &coords[12*n], &gradop[k*nelem*12+n*12], &det_j[npf*n+k], error, &lerr );
+
+      if ( lerr )
+        std::cout << "sorry, issue with face_grad_op.." << std::endl;
+    }
+  }
+}
+
+void threeD_gij(int npe, int nint,
+                const SharedMemView<DoubleType***>& deriv,
+                const SharedMemView<DoubleType**>& cordel,
+                SharedMemView<DoubleType***>& gupperij,
+                SharedMemView<DoubleType***>& glowerij)
+{
+  DoubleType dx_ds[3][3], ds_dx[3][3];
+  DoubleType det_j = 0.0, denom = 0.0;
+  const double realmin = 2.2250738585072014e-308;
+
+  for(int ki=0; ki<nint; ++ki) {
+    dx_ds[0][0] = 0.0;
+    dx_ds[0][1] = 0.0;
+    dx_ds[0][2] = 0.0;
+    dx_ds[1][0] = 0.0;
+    dx_ds[1][1] = 0.0;
+    dx_ds[1][2] = 0.0;
+    dx_ds[2][0] = 0.0;
+    dx_ds[2][1] = 0.0;
+    dx_ds[2][2] = 0.0;
+// 
+// calculate the jacobian at the integration station -
+    for(int kn=0; kn<npe; ++kn) {
+       dx_ds[0][0] += deriv(ki,kn,0)*cordel(kn,0);
+       dx_ds[1][0] += deriv(ki,kn,1)*cordel(kn,0);
+       dx_ds[2][0] += deriv(ki,kn,2)*cordel(kn,0);
+//                
+       dx_ds[0][1] += deriv(ki,kn,0)*cordel(kn,1);
+       dx_ds[1][1] += deriv(ki,kn,1)*cordel(kn,1);
+       dx_ds[2][1] += deriv(ki,kn,2)*cordel(kn,1);
+//                
+       dx_ds[0][2] += deriv(ki,kn,0)*cordel(kn,2);
+       dx_ds[1][2] += deriv(ki,kn,1)*cordel(kn,2);
+       dx_ds[2][2] += deriv(ki,kn,2)*cordel(kn,2);
+    }
+
+// calculate the determinate of the Jacobian at the integration station -
+    det_j= dx_ds[0][0]*(dx_ds[1][1]*dx_ds[2][2]-dx_ds[1][2]*dx_ds[2][1])
+         + dx_ds[0][1]*(dx_ds[1][2]*dx_ds[2][0]-dx_ds[1][0]*dx_ds[2][2])
+         + dx_ds[0][2]*(dx_ds[1][0]*dx_ds[2][1]-dx_ds[1][1]*dx_ds[2][0]);
+
+// clip
+    denom = stk::math::if_then_else( det_j <= 1.e+6*realmin, 1.0, 1.0/det_j );
+
+//std::cout<<"cpp: det_j "<<det_j<<" denom "<<denom<<std::endl;
+
+// calculate the inverse Jacobian
+    ds_dx[0][0]= denom*(dx_ds[1][1]*dx_ds[2][2]-dx_ds[1][2]*dx_ds[2][1]);
+    ds_dx[1][0]= denom*(dx_ds[1][2]*dx_ds[2][0]-dx_ds[1][0]*dx_ds[2][2]);
+    ds_dx[2][0]= denom*(dx_ds[1][0]*dx_ds[2][1]-dx_ds[1][1]*dx_ds[2][0]);
+//                                                                     
+    ds_dx[0][1]= denom*(dx_ds[0][2]*dx_ds[2][1]-dx_ds[0][1]*dx_ds[2][2]);
+    ds_dx[1][1]= denom*(dx_ds[0][0]*dx_ds[2][2]-dx_ds[0][2]*dx_ds[2][0]);
+    ds_dx[2][1]= denom*(dx_ds[0][1]*dx_ds[2][0]-dx_ds[0][0]*dx_ds[2][1]);
+//                                                                     
+    ds_dx[0][2]= denom*(dx_ds[0][1]*dx_ds[1][2]-dx_ds[0][2]*dx_ds[1][1]);
+    ds_dx[1][2]= denom*(dx_ds[0][2]*dx_ds[1][0]-dx_ds[0][0]*dx_ds[1][2]);
+    ds_dx[2][2]= denom*(dx_ds[0][0]*dx_ds[1][1]-dx_ds[0][1]*dx_ds[1][0]);
+//
+    for(int i=0; i<3; ++i) { 
+       for(int j=0; j<3; ++j) {
+          gupperij(ki,j,i) = 
+              dx_ds[0][i]*dx_ds[0][j]+dx_ds[1][i]*dx_ds[1][j]
+              +dx_ds[2][i]*dx_ds[2][j];
+//std::cout<<"gupperij("<<j<<","<<i<<"): "<<gupperij(ki,j,i)<<std::endl;
+         glowerij(ki,j,i) = 
+              ds_dx[i][0]*ds_dx[j][0]+ds_dx[i][1]*ds_dx[j][1]
+              +ds_dx[i][2]*ds_dx[j][2];
+       }
+    }
+  }
+}
+
+//--------------------------------------------------------------------------
+//-------- guij ------------------------------------------------------------
+//--------------------------------------------------------------------------
+void TetSCS::gij(
+    SharedMemView<DoubleType**>& coords,
+    SharedMemView<DoubleType***>& gupper,
+    SharedMemView<DoubleType***>& glower,
+    SharedMemView<DoubleType***>& deriv)
+{
+  threeD_gij(nodesPerElement_, numIntPoints_, deriv, coords, gupper, glower);
+}
+
+void TetSCS::gij(
+  const double *coords,
+  double *gupperij,
+  double *glowerij,
+  double *deriv)
+{
+  SIERRA_FORTRAN(threed_gij)
+    ( &nodesPerElement_,
+      &numIntPoints_,
+      deriv,
+      coords, gupperij, glowerij);
+}
+
+//--------------------------------------------------------------------------
+//-------- adjacentNodes ---------------------------------------------------
+//--------------------------------------------------------------------------
+const int *
+TetSCS::adjacentNodes()
+{
+  // define L/R mappings
+  return &lrscv_[0];
+}
+
+//--------------------------------------------------------------------------
+//-------- shape_fcn -------------------------------------------------------
+//--------------------------------------------------------------------------
+void
+TetSCS::shape_fcn(double *shpfc)
+{
+  tet_shape_fcn(numIntPoints_, &intgLoc_[0], shpfc);
+}
+
+//--------------------------------------------------------------------------
+//-------- shifted_shape_fcn -----------------------------------------------
+//--------------------------------------------------------------------------
+void
+TetSCS::shifted_shape_fcn(double *shpfc)
+{
+  tet_shape_fcn(numIntPoints_, &intgLocShift_[0], shpfc);
+}
+
+//--------------------------------------------------------------------------
+//-------- tet_shape_fcn ---------------------------------------------------
+//--------------------------------------------------------------------------
+void
+TetSCS::tet_shape_fcn(
+  const int  &npts,
+  const double *par_coord, 
+  double *shape_fcn)
+{
+  for (int j = 0; j < npts; ++j ) {
+    const int fourj = 4*j;
+    const int k = 3*j;
+    const double xi = par_coord[k];
+    const double eta = par_coord[k+1];
+    const double zeta = par_coord[k+2];
+    shape_fcn[fourj] = 1.0 - xi - eta - zeta;
+    shape_fcn[1 + fourj] = xi;
+    shape_fcn[2 + fourj] = eta;
+    shape_fcn[3 + fourj] = zeta;
+  }
+}
+
+//--------------------------------------------------------------------------
+//-------- opposingNodes --------------------------------------------------
+//--------------------------------------------------------------------------
+int
+TetSCS::opposingNodes(
+  const int ordinal,
+  const int node)
+{
+  return oppNode_[ordinal*3+node];
+}
+
+//--------------------------------------------------------------------------
+//-------- opposingFace --------------------------------------------------
+//--------------------------------------------------------------------------
+int
+TetSCS::opposingFace(
+  const int ordinal,
+  const int node)
+{
+  return oppFace_[ordinal*3+node];
+}
+
+//--------------------------------------------------------------------------
+//-------- isInElement -----------------------------------------------------
+//--------------------------------------------------------------------------
+double
+TetSCS::isInElement(
+    const double * elem_nodal_coor,
+    const double * point_coor,
+	  double * par_coor ) 
+{
+  // load up the element coordinates
+  const double x1 = elem_nodal_coor[0];
+  const double x2 = elem_nodal_coor[1];
+  const double x3 = elem_nodal_coor[2];
+  const double x4 = elem_nodal_coor[3];
+
+  const double y1 = elem_nodal_coor[4];
+  const double y2 = elem_nodal_coor[5];
+  const double y3 = elem_nodal_coor[6];
+  const double y4 = elem_nodal_coor[7];
+
+  const double z1 = elem_nodal_coor[8];
+  const double z2 = elem_nodal_coor[9];
+  const double z3 = elem_nodal_coor[10];
+  const double z4 = elem_nodal_coor[11];
+
+  // determinant of matrix M in eqn x-x1 = M*xi
+  const double det =
+    (x2 - x1)*( (y3 - y1)*(z4 - z1) - (y4 - y1)*(z3 - z1) ) -
+    (x3 - x1)*( (y2 - y1)*(z4 - z1) - (y4 - y1)*(z2 - z1) ) +
+    (x4 - x1)*( (y2 - y1)*(z3 - z1) - (y3 - y1)*(z2 - z1) );
+
+  const double invDet = 1.0/det;
+
+  // matrix entries in inverse of M
+
+  const double m11 = y3*z4 - y1*z4 - y4*z3+y1*z3+y4*z1 - y3*z1;
+  const double m12 =  - (x3*z4 - x1*z4 - x4*z3+x1*z3+x4*z1 - x3*z1);
+  const double m13 = x3*y4 - x1*y4 - x4*y3+x1*y3+x4*y1 - x3*y1;
+  const double m21 =  - (y2*z4 - y1*z4 - y4*z2+y1*z2+y4*z1 - y2*z1);
+  const double m22 = x2*z4 - x1*z4 - x4*z2+x1*z2+x4*z1 - x2*z1;
+  const double m23 =  - (x2*y4 - x1*y4 - x4*y2+x1*y2+x4*y1 - x2*y1);
+  const double m31 = y2*z3 - y1*z3 - y3*z2+y1*z2+y3*z1 - y2*z1;
+  const double m32 =  - (x2*z3 - x1*z3 - x3*z2+x1*z2+x3*z1 - x2*z1);
+  const double m33 = x2*y3 - x1*y3 - x3*y2+x1*y2+x3*y1 - x2*y1;
+
+  const double xx1 = point_coor[0] - x1;
+  const double yy1 = point_coor[1] - y1;
+  const double zz1 = point_coor[2] - z1;
+
+  // solve for parametric coordinates
+
+  const double xi   = invDet*(m11*xx1 + m12*yy1 + m13*zz1);
+  const double eta  = invDet*(m21*xx1 + m22*yy1 + m23*zz1);
+  const double zeta = invDet*(m31*xx1 + m32*yy1 + m33*zz1);
+
+  // if volume coordinates are negative, point is outside the tet
+
+  par_coor[0] = xi;
+  par_coor[1] = eta;
+  par_coor[2] = zeta;
+
+  std::vector<double> x(3);
+  x[0]=par_coor[0];
+  x[1]=par_coor[1];
+  x[2]=par_coor[2];
+
+  const double dist = parametric_distance(x);
+
+  return dist;
+}
+
+//--------------------------------------------------------------------------
+//-------- interpolatePoint ------------------------------------------------
+//--------------------------------------------------------------------------
+void
+TetSCS::interpolatePoint(
+    const int  & ncomp_field,
+    const double * par_coord,           // (3)
+    const double * field,               // (4,ncomp_field)
+	  double * result ) // (ncomp_field)
+{
+  const double xi   = par_coord[0];
+  const double eta  = par_coord[1];
+  const double zeta = par_coord[2];
+
+  const double psi1 = 1.0 - xi - eta - zeta;
+
+  for(int i = 0; i < ncomp_field; ++i) {
+    const int fourI = 4*i;
+
+    const double f1 = field[fourI];
+    const double f2 = field[1 + fourI];
+    const double f3 = field[2 + fourI];
+    const double f4 = field[3 + fourI];
+
+    result[i] = f1*psi1 + f2*xi + f3*eta + f4*zeta;
+  }
+}
+
+//--------------------------------------------------------------------------
+//-------- general_shape_fcn -----------------------------------------------
+//--------------------------------------------------------------------------
+void
+TetSCS::general_shape_fcn(
+  const int numIp,
+  const double *isoParCoord,
+  double *shpfc)
+{
+  tet_shape_fcn(numIp, &isoParCoord[0], shpfc);
+}
+
+//--------------------------------------------------------------------------
+//-------- general_face_grad_op --------------------------------------------
+//--------------------------------------------------------------------------
+void 
+TetSCS::general_face_grad_op(
+  const int /*face_ordinal*/,
+  const double */*isoParCoord*/,
+  const double *coords,
+  double *gradop,
+  double *det_j,
+  double *error)
+{
+  int lerr = 0;
+
+  const int nface = 1;
+  double dpsi[12];
+
+  // derivatives are constant
+  SIERRA_FORTRAN(tet_derivative)
+    ( &nface, dpsi );
+
+  SIERRA_FORTRAN(tet_gradient_operator)
+    ( &nface,
+      &nodesPerElement_,
+      &nface,
+      dpsi,
+      &coords[0], &gradop[0], &det_j[0], error, &lerr );
+  
+  if ( lerr )
+    throw std::runtime_error("TetSCS::general_face_grad_op issue");
+ 
+}
+
+//--------------------------------------------------------------------------
+//-------- sidePcoords_to_elemPcoords --------------------------------------
+//--------------------------------------------------------------------------
+void 
+TetSCS::sidePcoords_to_elemPcoords(
+  const int & side_ordinal,
+  const int & npoints,
+  const double *side_pcoords,
+  double *elem_pcoords)
+{
+  switch (side_ordinal) {
+  case 0:
+    for (int i=0; i<npoints; i++) {
+      elem_pcoords[i*3+0] = side_pcoords[2*i+0];
+      elem_pcoords[i*3+1] = 0.0;
+      elem_pcoords[i*3+2] = side_pcoords[2*i+1];
+    }
+    break;
+  case 1:
+    for (int i=0; i<npoints; i++) {
+      elem_pcoords[i*3+0] = 1.0 - side_pcoords[2*i+0] - side_pcoords[2*i+1];
+      elem_pcoords[i*3+1] = side_pcoords[2*i+0];
+      elem_pcoords[i*3+2] = side_pcoords[2*i+1];
+    }
+    break;
+  case 2:
+    for (int i=0; i<npoints; i++) {
+      elem_pcoords[i*3+0] = 0.0;
+      elem_pcoords[i*3+1] = side_pcoords[2*i+1];
+      elem_pcoords[i*3+2] = side_pcoords[2*i+0];
+    }
+    break;
+  case 3:
+    for (int i=0; i<npoints; i++) {
+      elem_pcoords[i*3+0] = side_pcoords[2*i+1];
+      elem_pcoords[i*3+1] = side_pcoords[2*i+0];
+      elem_pcoords[i*3+2] = 0.0;
+    }
+    break;
+  default:
+    throw std::runtime_error("TetSCS::sideMap invalid ordinal");
+  }
+  return;
+}
+
+//--------------------------------------------------------------------------
+//-------- parametric_distance ---------------------------------------------
+//--------------------------------------------------------------------------
+double
+TetSCS::parametric_distance(const std::vector<double> &x)
+{
+  const double X=x[0] - 1./4.;
+  const double Y=x[1] - 1./4.;
+  const double Z=x[2] - 1./4.;
+  const double dist0 = -4*X;
+  const double dist1 = -4*Y;
+  const double dist2 = -4*Z;
+  const double dist3 =  4*(X+Y+Z);
+  const double dist  = std::max(std::max(dist0,dist1),std::max(dist2,dist3));
+  return dist;
+}
+
+} // namespace nalu
+} // namespace sierra
+
